moved ruid handling into tasks
