+++ packages/deprecation-crawler/src/lib/tasks/crawl.ts
@@ -1,7 +1,7 @@
 import { CrawlConfig, CrawledRelease, CrawlerProcess } from '../models';
 import { concat, tap } from '../utils';
 import { crawlDeprecations, getSourceFiles } from '../crawler';
-import { addRuid } from '../processors/add-ruid';
+import { addRuid } from './add-ruid';
 import {
   printFooterLine,
   printHeadline,
